Ensure the indicator span timer name is un-prefixed (#484)
Due to the NamePrefix global, all SSF metrics generated by veneur
would be prefixed by "veneur.". This is quite undesirable for the
metrics that constitute the public interface for veneur! This change
undoes the prefixing by setting the name property on that metric
before parsing it into an internal metric.