Final gRPC integration with veneur-proxy (#524)
* Final gRPC integration with veneur-proxy

This change is the final piece to enabling full support for gRPC metrics
forwarding.  It pulls in the "proxysrv" package to the main Proxy, and
adds a separate Consul service and destination list to use with it. With
this addition, I was able to add in the final large integration test
that verifies all of this behavior.

I did make some breaking changes to the behavior of the Consul
discoverer.  Currently it returns destinations with an protocol prefix
in the form "http://<host>:<port", which isn't suitable for gRPC, as it
obviously doesn't want the prefix.  I modified the Consul discoverer to
instead just return strings of the form "<host>:<port>" without the
prefix.

Changes were also necessary to the main Proxy type to support two
listeners (HTTP and gRPC).  Similar to the changes in #454, I added
appropriate methods for starting and stopping the gRPC server
specifically, and then edited (*Proxy).Serve() to start one or both of
the listeners.  The function should then block until either of them
exits.

Testing Done:
* Test that the full gRPC forwarding pipeline works: I added an
end-to-end test consisting of a local Veneur, a proxy, and finally a
global instance all communicating locally over gRPC!
* Test that the changes to the Consul discoverer aren't breaking:
    * I fixed the Consul discoverer's test cases for the new behavior.
    * I checked that tests failed without including the HTTP prefix
    elsewhere, then that they passed once I added a change to set the
    prefix.
* Test that starting and stopping the gRPC server works correctly:
    * I added a test that verifies that the main (*Proxy).Serve() method exits
    when the gRPC server (only) is stopped.
    * I also added a similar test for HTTP: that the (*Proxy).Serve()
    method quits when just the HTTP server is stopped.

* proxy.go: fix spelling errors

* Revert removing proxy runtime metric goroutine

I accidentally removed a goroutine which emits runtime metrics for the
proxy.  This change just adds that back in.

* Close channels rather than send empty structs

As suggested in PR comments, rather than sending empty structs as
sentinel values we can instead just close the channel! This is useful
in multiple areas of the changes to add full proxy support.

* Use a buffer size of 2 when closing the proxy

The method that handles closing the different proxy listeners (gRPC and
HTTP) creates a channel that can be sent to by two different goroutines.
We might as well give it a buffer of size 2 so that sends to the channel
can never block.

* Revert changes to generate gRPC protobufs

I realized that I never added the proper commands to
generate.go to regenerate the Golang protobuf files from the
definitions.  As suggested by @sjung-stripe, rather than including them
in this change, I can just put it in a separate PR.

* Add a CHANGELOG entry