Fix flaky gRPC forward tests
The TestProxyServeStopHTTP goroutine waits for the atomic integer to be
flipped from the server goroutine, and then sends a graceful shutdown
signal. This might happen before the server goroutine actually
instantiates the graceful listener. Then the listener would never
receive the shutdown, and would keep listening forever. Instead, the
graceful listener should be instantiated before the atomic integer is
flipped.

In the forwarding test fixtures, the proxy server spins up immediately
after the global server, and might attempt to forward to the global
before the global's ports are actually opened. Fixture creation should
block until all the servers are started.