Automate code generation for gRPC forwarding
In an earlier PR (#439) I added in a few different protobuf and gRPC
definitions, as well as the generated packages for each of them.  I had
generated them locally, but I didn't add the proper "protoc" invocations
to "generate.go" to make it easy to regenerate these in the future.

This change adds the proper commands to the Travis build and
"generate.go" to allow you to run "go generate" both locally and in the
CI environment.  Specifically it adds the "protoc-gen-gogofaster"
binary, which generates code from protobuf definitions that is supposed
to be a bit leaner than "protoc-gen-gofast".  There isn't any particular
reason that we need to use "gogofaster", it was just what I used to
originally generate all the protobuf-related code, so changing it would
generate a large diff.

I'm happy to change it back to the normal "protoc-gen-gofast" if that is
preferable.