Add sink for sending spans to Splunk (#531)
* Add stub for SplunkSpanSink

* Flesh out the splunk SpanSink implemenation, vendor HEC

* Use the correct version of uuid

* Use configurable HEC tokens and HTTP addresses

Now veneur compiles again & we can pass the config values through!

* FTTB, don't validate tls certs on the HEC side

Our splunkindexers can't present a cert that usefully identifies them
to clients, so we have to turn off validation ):

* Specify structure of Splunk event message output

* Allow configuring a server name to TLS-verify

This is useful for those situations where your splunk HEC ingesters
don't present a name that differs from the name you have to use on the
connection URL.

* Make start_timestamp and end_timestamp time.Time

This should print nicer timestamps now!

* Set the span's start timestamp as the event timestamp

This way, we get them in the order that they were started, not ended;
I'm not sure that's the right way to do it, but it feels... righter to
me? Let's see how it looks!

* Set hostname and sourcetype on the HEC event

Hostname should be the host on which the span got reported, and
sourcetype should probably be the service. Maybe we can make this
configurable, but these seem reasonable for now.

* Try using conf.Hostname?

* Add logging and span submission success/failure tracking to HEC sink

* Oops, set the logger

* Correctly initialize the splunk span sink

I'd messed up the order of hostnames. Oops!

* Only log errors for span submission when the context deadline is fine

* Make a test case for the splunk span sink

This just tests that the splunk events we report come with the right
data & metadata.

* Fix time metadata setting on splunk HEC events

We only set the whole number of seconds correctly, which stripped
milliseconds out. Let's not do that instead.

* Changelog entry for HEC sink

* Make the HEC send timeout configurable

If unset or 0, there's no timeout.

* Represent HEC event timestamps as unix timestamps

They are now seconds since the UNIX Epoch. That's a format that splunk
can deal with better (and the timestamp on an event is readable, too).

* Use fewer useless temporary values