Reimplement the splunk HEC sink as a streaming HTTP client
This change drops the splunk-hec-go library and implements the sink as
a streaming HTTP client (with an io.Pipe that gets closed & recreated
once <limit> spans have been consumed).

This gets rid of all the implicit behavior in splunk-hec-go that we
don't want (automatic retries, backoffs, chunking, etc) and allows us
to focus on doing the right thing with a smaller interface.