Improve testability of the splunk span sink
This adds a second interface, TestableSplunkSpanSink that the sink
implements, which sticks a bunch of helpful methods on, for
synchronizing and ensuring all the events get sent when tests expect
them.

Also, fix an issue in tests where events can get delivered
out-of-order; it's an ok thing to happen, so make the test less
strict.