Correctly construct signalfx endpoint URLs
Since signalfx's ingestion endpoint is now less tolerant of URLs that
start with two /es, we should make sure that we don't accidentally end
up with those if someone should specify a base URL ending with a slash
in their config.

The best way to do that is to use go's built-in URL resolution
functionality. Let's just use that. This change also ensures that no
invalid URLs can be set as ingestion endpoints - veneur would panic at
startup if someone tried that.