Make Server.Flush synchronous
This helps us clear up the context-cancelation story in Server.Flush,
and it makes the function more clearly testable. Previously, it
spawned goroutines that terminated "sometime later" (and hopefully
before the next flush). Now, we wait for the forward to finish before
returning from Flush; context cancelation should ensure that veneur
really doesn't get stuck anymore.

Also, cancel any in-flight flush context as soon as we detect that the
server is terminating.