Clear up the Server.Flush method some more (#717)
* Make Server.Flush synchronous

This helps us clear up the context-cancelation story in Server.Flush,
and it makes the function more clearly testable. Previously, it
spawned goroutines that terminated "sometime later" (and hopefully
before the next flush). Now, we wait for the forward to finish before
returning from Flush; context cancelation should ensure that veneur
really doesn't get stuck anymore.

Also, cancel any in-flight flush context as soon as we detect that the
server is terminating.

* Fix a test timeout around flushing

This test called Flush, expecting the flush to not wait; it does wait
however, so let's rely on the running server's flush. To ensure the
test doesn't terminate before flush is done, wait until we receive a
local metric.

* Make the "unreachable destinations" test abort sooner

This should reduce the number of times this test times out locally a
bit.