Restructure the way we submit per-api key batches
This function had a potential hang: If a batch was larger than the
"max submission" size, the function deadlocked. Instead of reading the
errors (and not doing a good job of it) so that we could set
span.Error, use a sync.Once and pass the errors through to a function
that aggregates them and return that aggregate.