Fix some flaky tests (#725)
* Fix flaky TestE2EForwardingGRPCMetrics test

It would sometimes not get all the metrics it expected, and so flake
out. This change should ensure we keep reading metrics until all the
expected ones (and no unexpected ones) came through.

* Setup and use a debug-level logger in splunk sink tests

* Debug-log the entire response body if something should go wrong

Let's only do this if we're logging at debug level, but the field is
valuable in order to track down the flaky annoying test failure.

* Do not set example.com as a HEC endpoint

This chunders a ton of 404s and muddles up other problems.