Wait for forwarding subconns to be ready
Normally, when a gRPC channel is in a TRANSIENT_FAILURE state (eg all
the underlying connections are reset), it's supposed to fail immediately
and escalate the problem to the caller. This option causes the gRPC
client to wait for the channel to return to the READY state and then try
the failed message again. Newer versions of the gRPC library actually
call this WaitForReady, but we're vendoring an older version that only
has the deprecated FailFast. More details:

https://github.com/grpc/grpc/blob/master/doc/connectivity-semantics-and-api.md
https://github.com/grpc/grpc/blob/master/doc/wait-for-ready.md