fixes https://github.com/stripe/veneur/issues/588
Currently when you try to store data using the s3 plugin, it will
always fail with the error

```"Error posting to s3" error="BucketRegionError: incorrect region, the bucket is not in 'us-west-2' region\n\tstatus code: 301, request id: , host id: " metrics=2```

This error is misleading as the real issue is that the bucket name is never set in the code.
The code is trying to use the Global Constant S3Bucket when it should be using the method receivers struct
`p *S3Plugin` and accessing the field `p.S3Bucket` instead of the constant S3Bucket that isn't
referenced anywhere in the code.