Add initial functionality to extend metrics tags
The previous config for "tags" is passed to sinks and added by sinks.
This causes any tags thus specified to be excluded from flushes to
global Veneur instances. In order to support flushing implicit tags
to global Veneurs, the least-worst place to put the code is in the
parsing logic, since that "touches" everything and is mostly independent
of the source of the data.

I chose to turn the parse functions into receivers on a struct; the struct
only contains "ExtendTag" for now. This allows us to take the config where
we have access to it (NewServer) and bind it to the functions that perform
message parsing (ParseMetric, ParseMetricSSF, etc). These functions then
pass both the tags they've derived and the new "ExtendTag" struct into
UDPMetric.UpdateTags, which centralizes the merging, sorting, joining,
and digest calculation of tags for a given UDPMetric.

There are no guards yet against mutating those fields in place; this
change set is already going to be pretty big.