Optimize ExtendTags for parity with before-code
This adds and tests a separate "shouldDropPrefix" method, which is
responsible for excluding tags that get overridden by implicit tags.
This uses a naive for-loop with the expectation that the avoidance
of strings.Split and key hashing will win out for small implicit
tag list sizes, which is the expected usage.

Also update the parser_test benchmark which had a couple typos,
and adds runs that include implicit tags to the statsd parser
benchmark (including one that sometimes overlaps)