Switch to pointer receivers for everything
I had a red herring where it looked like this branch increased the RSS
of a memory-heavy instance in QA. However, the increase remained after
rolling back to master. I had prepared this branch to test with in the
hope that it would reduce memory usage; it might have, a little, but I
did notice that it reduced CPU, so I'm keeping the changes.

Also added a benchmark against ConvertIndicatorMetrics which shows
significant differences with master, still investigating.