Define tags in config that can be flushed to global Veneur instances (#832)
* Add initial functionality to extend metrics tags

The previous config for "tags" is passed to sinks and added by sinks.
This causes any tags thus specified to be excluded from flushes to
global Veneur instances. In order to support flushing implicit tags
to global Veneurs, the least-worst place to put the code is in the
parsing logic, since that "touches" everything and is mostly independent
of the source of the data.

I chose to turn the parse functions into receivers on a struct; the struct
only contains "ExtendTag" for now. This allows us to take the config where
we have access to it (NewServer) and bind it to the functions that perform
message parsing (ParseMetric, ParseMetricSSF, etc). These functions then
pass both the tags they've derived and the new "ExtendTag" struct into
UDPMetric.UpdateTags, which centralizes the merging, sorting, joining,
and digest calculation of tags for a given UDPMetric.

There are no guards yet against mutating those fields in place; this
change set is already going to be pretty big.

* Update tests to exercise ExtendTag behavior

* Optimize ExtendTags for parity with before-code

This adds and tests a separate "shouldDropPrefix" method, which is
responsible for excluding tags that get overridden by implicit tags.
This uses a naive for-loop with the expectation that the avoidance
of strings.Split and key hashing will win out for small implicit
tag list sizes, which is the expected usage.

Also update the parser_test benchmark which had a couple typos,
and adds runs that include implicit tags to the statsd parser
benchmark (including one that sometimes overlaps)

* Switch to pointer receivers for everything

I had a red herring where it looked like this branch increased the RSS
of a memory-heavy instance in QA. However, the increase remained after
rolling back to master. I had prepared this branch to test with in the
hope that it would reduce memory usage; it might have, a little, but I
did notice that it reduced CPU, so I'm keeping the changes.

Also added a benchmark against ConvertIndicatorMetrics which shows
significant differences with master, still investigating.

* Move the benchmark with related benchmarks

I don't know why a bunch of samplers-package related tests are in
that root file...

* Rename dropPrefixes to avoid confusing naming. Elaborate on behavior in comments.

* Add explicit dependency on golang.org/x/net for veneur-emit (#833)