Upgrading prometheus to v0.42.0
We currently use a retracted version of Prometheus, which is missing several
nice testing utilities. The protobuf of the old version is compatible with, but
different from, the versions used by Cortex and Amazon Managed Prometheus. This
commit upgrades the version and makes some changes to use the newer generated
proto definitions.