Bugfix: Specifying multiple tags/labels with -a causes sporadic incorrect metric calculations
This fixes a bug that caused diff calculations to sporadically break when multiple labels
were added via the -a flag.  Labels were stored as key/values in a map and returned in a potentially different order
each time since map iteration order is random.  This caused the translator to emit the cumulative value instead of
the diffed value, because it would seem as though the metric was new when in fact it had been previously cached
with a different label ordering.  For example, the translator might look for "counter-key2:value2-key1:value1" when
it had been previously cached as "counter-key1:value1-key2:value2"